### 2022-03-07 #4229 @editorialbot
[PRE REVIEW]: CESAR-P: A dynamic urban building energy simulation tool
**Submitting author:** @kristina-o (<a href="http://orcid.org/0000-0001-6491-7641">Kristina Orehounig</a>)
**Repository:** https://github.com/hues-platform/cesar-p-core
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** V2.0.1
