### 2022-03-07 #4229 @Kevin-Mattheus-Moerman
@kristina-o thanks for this submission. I am the AEiC on call this week and have started to look for a handling editor. In the meantime, can you try to address to potentially invalid DOIs? :point_up: You can call `@editorialbot check references` here to check the references and use `@editorialbot generate pdf` to update the paper as well. Thanks.
