### 2022-03-07 #4229 @LeonieFierz
@Kevin-Mattheus-Moerman thank you for looking into our submission. I tried to fix the invalid DOIs, but could not get this one
"10.1016/S0378-7788(00)00114-6" working because the brackets are removed. Could you please tell me how to escape them? I tried with (( and \( and changed back to just the single bracket now.
