### 2022-03-08 #4229 @timtroendle
> @editorialbot invite @timtroendle as editor

@openjournals/joss-eics, happy to serve as an editor for this submission.
