### 2022-03-10 2e7ad54
Squashed commit of version 2.2.0:
