### 2022-03-21 #4229 @timtroendle
👋 @nmstreethran & @willu47, would any of you be willing to review this submission for JOSS?
