### 2022-04-25 #4261 @timtroendle
Thanks for the update, @noah80. You can simply leave this box unchecked for now.

We don't require contributions of authors to be code contributions. @kristina-o, can you clarify this point?
