### 2022-04-27 #4261 @kristina-o
Thank you @noah80 for reviewing our submission and @timtroendle for forwarding the question. My contributions to CESAR-P (and also the predecessor CESAR) include the original idea for the tool, contributing to the development of the approach, and supervision of further developments. The current version of CESAR-P was coded by our software engineer Leonie Fierz, she is listed together with me as the main authors. Hope that clarifies my contributions.
