### 2022-04-30 #4261 @jasondegraw
The software described in the paper is intended for modeling and simulation of sets of buildings (urban areas, campuses, etc.) and is implemented in Python. It is well documented and (at least in the files that I looked at) well written. The paper describes the need and features of the software, but I would not recommend acceptance of the paper without revisions. I do not believe these revisions are major revisions. The paper revisions are:

* The paper (and the documentation as well) is missing a clear statement of the intended audience. This is particularly important as urban BEM becomes more and more popular. This is why both of the "statement of need" items on the checklist are unchecked.
* The paper does not clearly address the state of the field. There are other options available and this one looks like it does have some unique features, but that isn't clear from the paper.
* The paper does not have enough about what the software can do. It would be much better if an example analysis was presented (in abbreviated form, perhaps one of the examples).

Other items from the checklist:

* The installation documentation works well and there are examples, but there are some minor typos and the simple example fails initially. On the bright side, the error clearly states that there needs to be a TEMP environment variable defined, and doing that allows the example to run. Maybe I missed a step in the directions, but it would be good to make sure that the example runs without modifications. That example also generates a fair number of (probably harmless) deprecation warnings from pandas. I also looked into the EnergyPlus outputs a little, and I did see a warning about unused objects. Those are also probably harmless, but speaking from experience (with EnergyPlus), some users can be put off by even harmless warnings.
* There are tests, but I was unable to run them. There's a lot of documentation, so perhaps I just missed that.
* There is developer documentation, but I did not find information on how to contribute or get support. Another thing that I maybe just missed.

One thing not really in the checklist:

* I did not find a clear statement on supported platforms. It looks like it is probably any platform that E+ runs on that supports Python, but it would be better if there was a clear statement on where it should be expected to work. I ran it on a fresh instance of Ubuntu 20.04 and it worked well as far as I could tell.

In summary, there are some things I'd like to see fixed, but in general @kristina-o and coworkers are to be congratulated on this work. I'm happy to follow up on any of the above items by filing issues or whatever else would be desired.
