### 2022-05-02 #4261 @LeonieFierz
@jasondegraw thank you for the review and the constructive feedback! We will go through it in detail and improve the points you mentioned.
Just two points from the technical side:

- The tests are run in a CI pipeline on our internal GitLab Server and it should actually be possible to run them with pytest after installing the package for development with poetry.
- I will think about how we can improve regarding the missing TEMP environment variable, which actually is a default on Windows.
