### 2022-05-12 #4261 @LeonieFierz
@jasondegraw That would be great if you could try to run the tests. To do so, you first need the "development installation" as described in https://cesar-p-core.readthedocs.io/en/latest/development/development-installation.html. With that set up, you can start the tests e.g. with

`poetry run pytest tests`

Note that some of the tests need EnergyPlus 9.5.0 installed and configured in the environment variables ENERGYPLUS_EXE and ENERGYPLUS_VER (see https://cesar-p-core.readthedocs.io/en/latest/installation.html#energyplus).
Tests depending on E+ Installation: tests\test_eplus_adapter\test_eplus_adapter.py, tests\test_eplus_adapter\test_solar_potential.py, tests\test_integration\test_cesar_integration.py, tests\test_model\test_model_pickling.py

And you mentioned that there were a few typos you had to fix in the cesar-p-usage-examples/simple_example. I just tested it again and it run without any adaption on Windows. Do you remember what was the problem if there was any besides the missing TEMP environment variable?
