### 2022-05-23 #4261 @noah80
I've installed the tool a couple of weeks back, but I stalled after that. I'm still interested in the review, but I would suggest a number of improvements to the authors:

- I've found the installation instructions here rather confusing:
https://cesar-p-core.readthedocs.io/en/latest/development/development-installation.html
https://cesar-p-core.readthedocs.io/en/latest/installation.html#installation

- It's not clearly described when I would need the development version and for which use the I want the normal install
- Unclear if it needs to be exactly Python 3.8 or only newer than 3.8
- There could be a batch file that checks if the install succeeded and does everything automatically, but instead I have a long list of things I might want to check
- After installing the core library it's completely unclear how to get a first example running and how to play with that.

Do these suggestions make sense? If so, I'd suggest implementing them and that would make it easier for the reviewers to actually get an example running to review the results. The software sounds pretty amazing after all.
