### 2022-05-25 #4261 @LeonieFierz
@noah80 thank you for starting to review our submission

we will try to state the points you mentioned in more detail.

meanwhile to get you started with playing with the library please have a look on https://github.com/hues-platform/cesar-p-usage-examples where you find different usage examples - please let me know if you have troubles with that (if you are running it on a linux system, please set an environment variable named TEMP pointing to location where temporary files can be created - we fixed that issue already but it is not yet pushed to github)
