### 2022-06-03 hues-platform/cesar-p-core#2 @olejandro
JOSS Software Paper: Statement of Need
- [x] Target audience of the software is not stated. It seems to be implied that the audience is limited to users of EnergyPlus, but is it really so?
- [x] It is stated that the software addresses the need of generating various profiles which are often not readily available. However it is not clear why some of the identified key features are important, e. g.: use of geo-referenced data sources as input for building geometries or storing archetypes in a graph database. Please explain importance of these features.

openjournals/joss-reviews#4261
