### 2022-06-03 hues-platform/cesar-p-core#3 @olejandro
JOSS Software Paper: State of the Field
- [ ] Please describe how this software compares to other commonly-used packages.

openjournals/joss-reviews#4261
