### 2022-06-03 hues-platform/cesar-p-core#4 @olejandro
JOSS Software Paper: Quality of Writing
- [ ] Please edit the paper for language. E.g.: _"~~The~~ modellling of indoor temperatures, ..."_

openjournals/joss-reviews#4261
