### 2022-06-03 hues-platform/cesar-p-core#5 @olejandro
JOSS Software Paper: References
- [ ] Please update the reference list based on #3

openjournals/joss-reviews#4261
