### 2022-06-03 hues-platform/cesar-p-core#6 @olejandro
JOSS Functionality: Some Tests Fail
FAILED tests/test_model/test_model_pickling.py::test_model_serialize - eppy.modeleditor.IDDAlreadySetError: IDD file

openjournals/joss-reviews#4261
