### 2022-06-03 hues-platform/cesar-p-core#7 @olejandro
JOSS Documentation: Statement of Need
The documentation does not seem to state:
- [ ] Target audience

openjournals/joss-reviews#4261
