### 2022-06-03 hues-platform/cesar-p-core#8 @olejandro
JOSS Documentation: Community Guidelines
Please include e.g. in the readme guidelines for third parties wishing to:

- [x] contribute to software

openjournals/joss-reviews#4261
