### 2022-06-26 hues-platform/cesar-p-core#6 @LeonieFierz
Could you please give some more details about the envirnoment you are trying to run the tests, respectively to try again running with E+ 9.5.0 if you did not use this E+ Version:

- EnergyPlus Version (should be 9.5.0 to run tests successfully, I will add this information to the development docu)
- Did you set ENERGYPLUS_VER and ENERGYPLUS_EXE environment variables? If not please them to the installation location of EnergyPlus 9.0.5, as described in https://cesar-p-core.readthedocs.io/en/latest/installation.html#energyplus
- Python Version
- What command did you use to run the tests? Running test over the IDE might fail becasue for some I need to switch the used IDD and I remember having problems when running through the IDE because it does not start a clean instance for each of the tests which is necessary because EPPY is using a singelton instance not allowing to change E+ Version resp the used IDD

see also https://github.com/openjournals/joss-reviews/issues/4261#issuecomment-1125403565
