### 2022-07-03 #4261 @kristina-o
> Dear @olejandro please find below the contributions of the other 3 remaining authors:
•	James Allan has developed and integrated the Graph Database data for Archetypical constructions
•	Sven Eggimann developed a shapefile parser for reading site vertices
•	Natasa Vulic contributed with development testing and support and performed a validation study at national scale.
These three parts are important for the functioning of the tool. Additionally all three of them (as well as Leonie Fierz and myself) supported the open access release of CESAR-P, the preparation of the JOSS paper publication and are actively contributing to the further development and testing of CESAR-P. Ricardo Silva has developed an optional feature for CESAR-P but didn't further contribute to the development of the tool, and the main contributions of Aaron Bojarski are not yet implemented in the current open-source version of CESAR-P. That’s why the latter two are not included in the list of authors. I hope this clarifies our motivation for defining the list of authors.
