+++ paper/paper.md
@@ -1,0 +1,339 @@
+---
+title: 'CESAR-P: A dynamic urban building energy simulation tool'
+tags:
+  - Python
+  - buildings
+  - energy consumption
+  - cities
+  - retrofitting solutions
+  - costs and emissions
+authors:
+  - name: Kristina Orehounig^[co-first author] ^[corresponding author] 
+    orcid: 0000-0001-6491-7641
+    affiliation: 1
+  - name: Leonie Fierz^[co-first author]
+    orcid: 0000-0002-1457-537X
+    affiliation: 1
+  - name: James Allan
+    affiliation: 1
+  - name: Sven Eggimann
+    affiliation: 1 
+  - name: Natasa Vulic
+    affiliation: 1 
+affiliations:
+  - name: Urban Energy Systems Laboratory, Empa, Duebendorf, Switzerland
+    index: 1
+date: 10 July 2022
+bibliography: paper.bib
+---
+
+# Summary 
+
+Buildings are responsible for a large share of energy consumption and carbon dioxide emissions. With the challenge of
+reducing energy consumption and integrating renewable energy sources in buildings and neighbourhoods, an understanding
+of energy consumption patterns is needed at different temporal and spatial scales. CESAR-P (Combined Energy Simulation
+And Retrofitting - Python) is a bottom-up building stock modelling software to evaluate energy consumption and
+retrofitting strategies for individual buildings and neighbourhoods. CESAR-P is used to parameterize models for the
+dynamic building energy simulation tool EnergyPlus (@CRAWLEY_ET_AL_2001). It computes hourly energy demand profiles and
+indoor temperatures at the building level and takes into account shading and reflections of surrounding buildings. After
+the current energy demand is computed, retrofitting measures for individual buildings can be evaluated in terms of
+energy savings, embodied emissions and costs. The default library of CESAR-P is based on Swiss archetypal buildings
+created from standards and statistics about the Swiss building stock; however, users are able to define their own
+constructions and internal conditions for their buildings. CESAR-P is written in Python and is a further development of
+the tool CESAR (@WANG_ET_AL_2018). The building model generator of CESAR-P creates EnergyPlus input data files (.IDF)
+for each building within a specified extent, and can process geo-referenced input data as shapefiles (ESRI .shp format)
+for geometry. Building usage type and construction information is used to populate the building model. The EnergyPlus
+files are executed with a weather file (EnergyPlus epw weather data format) of the geographic location which provides
+the necessary climatic context for the building. The results consist of a configurable set of yearly demand values and
+hourly resolved times series for each building from EnergyPlus and additional operational costs and emissions. The
+retrofit module can be used to update the building models according to a specified retrofit strategy. Retrofit
+strategies specify the frequency of retrofits to the main building elements such as walls, windows and roofs across the
+building stock. The output of the retrofit module include, in addition to the above-mentioned operational indicators,
+costs and embodied emissions of retrofitting interventions. All simulation steps can also be run in parallel on multiple
+cores. The code is platform-independent and is tested to run on Linux and Windows. The source code for CESAR-P has been
+archived to Zenodo: [@cesar-p-zenodo] 
+
+# Statement of need and Key features 
+
+Building energy simulation is used to generate the expected energy demand profiles of buildings. Building energy demand
+profiles are useful for planners to evaluate current performance and investigate how buildings will perform under
+different scenarios (e.g. building retrofits, changing climate). Measured demands are often unavailable which makes
+building energy simulation a key tool for decision-making. Urban building energy modelling (UBEM) has been established
[…]
