+++ paper/paper.bib
@@ -1,0 +1,97 @@
+@article{WANG_ET_AL_2018,
+  title = {CESAR: A bottom-up building stock modelling tool for Swit-zerland to address sustainable energy transformation strategies},
+  journal = {Energy and Buildings},
+  volume = {169},
+  pages = {9-26},
+  year = {2018},
+  issn = {0378-7788},
+  doi = {10.1016/j.enbuild.2018.03.020},
+  url = {https://www.sciencedirect.com/science/article/pii/S0378778817337696},
+  author = {Danhong Wang and Jonas Landolt and Georgios Mavromatidis and Kristina Orehounig and Jan Carmeliet},
+  keywords = {Building stock model, Dynamic building simulation, Neigh-bourhood, Retrofitting, Climate change}
+}
+
+@article{CRAWLEY_ET_AL_2001,
+  title = {EnergyPlus: creating a new-generation building energy simulation program},
+  journal = {Energy and Buildings},
+  volume = {33},
+  number = {4},
+  pages = {319-331},
+  year = {2001},
+  note = {Special Issue: BUILDING SIMULATION'99},
+  issn = {0378-7788},
+  doi = {10.1016/S0378-7788(00)00114-6},
+  url = {https://www.sciencedirect.com/science/article/pii/S0378778800001146},
+  author = {Drury B. Crawley and Linda K. Lawrie and Frederick C. Winkelmann and W.F. Buhl and Y.Joe Huang and Curtis O. Pedersen and Richard K. Strand and Richard J. Liesen and Daniel E. Fisher and Michael J. Witte and Jason Glazer},
+  keywords = {Building performance, Simulation, Energy performance, Heat balance, Mass balance, Modular simulation},
+}
+
+@article{EGGIMANN_ET_AL_2022,
+  title = {Spatiotemporal upscaling errors of building stock clustering for energy demand simulation},
+  journal = {Energy and Buildings},
+  volume = {},
+  pages = {111844},
+  year = {2022},
+  issn = {0378-7788},
+  doi = {10.1016/j.enbuild.2022.111844},
+  url = {https://www.sciencedirect.com/science/article/pii/S0378778822000159},
+  author = {Sven Eggimann and Natasa Vulic and Martin Rüdisüli and Robin Mutschler and Kristina Orehounig and Matthias Sulzer},
+  keywords = {}
+}
+
+@article{SILVA_ET_AL_2022,
+  title = {Opportunities for passive cooling to mitigate the impact of climate change in Switzerland},
+  journal = {Buildings and environment},
+  volume = {},
+  pages = {108574},
+  year = {2022},
+  issn = {0360-1323},
+  doi = {10.1016/j.buildenv.2021.108574},
+  url = {https://doi.org/10.1016/j.buildenv.2021.108574},
+  author = {Ricardo Silva and Sven Eggimann and Léonie Fierz and Massimo Fiorentini and Kristina Orehounig and Luca Baldini},
+  keywords = {}
+}
+
+@software{cesar-p-zenodo,
+  author = {Léonie Fierz and UrbanEnergySystemsLab},
+  title = {hues-platform/cesar-p-core},
+  month = {apr},
+  year = {2021},
+  publisher = {Zenodo},
+  doi = {10.5281/zenodo.4682880},
+  url = {https://doi.org/10.5281/zenodo.4682880}
+}
+
+@INPROCEEDINGS{ALLAN_ET_AL_2021,
+  author = {James Allan and Leonie Fierz and Andrew Bollinger and Kristina Orehounig},
+  title = {Linked Data Ontology for Urban and National Scale Building Energy Modelling},
+  booktitle = {eSim conference proceedings},
+  year = {2021}
+}
+
[…]
