### 2022-07-14 hues-platform/cesar-p-core#8 @LeonieFierz
Contribution section was added to README in cesar-p-core as well as in cesar-p-usage-examples
