### 2022-07-14 hues-platform/cesar-p-core#3 @LeonieFierz
a short description about the state of the field has been added to section "Statement of need and Key feauters"; published with release 2.3.0
