### 2022-07-14 hues-platform/cesar-p-core#2 @LeonieFierz
Mentioned points were added to "Statement of need and Key features" section; published with release 2.3.0
