### 2022-07-14 #4261 @LeonieFierz
@noah80 thank you again for looking into cesar-p-core. Please note following updates with version 2.3.0:

- Installation instructions have been improved, for cesar-p-core and cesar-p-usage-examples. See https://cesar-p-core.readthedocs.io/en/latest/installation.html to get started. It links to cesar-p-usage-examples to get you started after the installation.
- As the default installation is straight forward (and with the updated instructions easier to follow) a batch script to check the installation should not be necessary
- Updated the README to include only Python 3.8 as supported version, as we only run the CI with Python 3.8. Even though it most probably will run with newer Python Versions, but this is up to the user to check.
