### 2022-07-14 #4261 @LeonieFierz
@olejandro thank you for your review! we addressed the points you created issues for with version 2.3.0. I closed the issues and mentioned what we did shortly in the issue comment.
