### 2022-09-13 #4261 @noah80
I've got it to run and execute the simple example. Everything seems to be working.

The updated paper is much nicer, thanks for that! Except it seems to be in the wrong branch? As far as I can see the updated paper is in the main branch and the joss-branch has the old version.

The only thing I haven't found is the "State of the field: Do the authors describe how this software compares to other commonly-used packages?". So that's the last unchecked item for me.

Other than that, very nice work!

PS: For me a couple of unit tests seem to be failing. But that could just be my installation. Maybe set up a github action sometime to run the tests automatically?
