### 2022-09-18 #4261 @LeonieFierz
@noah80 Thank you for your feedback.

We pushed the latest version including the JOSS Paper to the master, I will delete the joss-paper branch once the Review process has finished.

The ""State of the field: Do the authors describe how this software compares to other commonly-used packages?" is briefly addressed with ".... Various review papers summarize available tools, including their key-features and limitations [@JOHARI_ET_AL_2020]. The here presented tool CESAR-P falls within the category of bottom-up modelling approaches. ...".
Does this fulfill the point for you or does it need some more details?

p.s. the integration test relying on EnergyPlus need the setup for the tests as descriped in the documentation, so even if the example work it might well be that those integration tests fail. We develop and have the testing set up internally on GitLab where we have continious integration set up.
