### 2022-09-29 #4261 @noah80
Fine for me.

Just wanted to point out that the bot seems to be configured to take the paper from the JOSS branch, which is completely outdated.
