### 2022-09-29 #4261 @timtroendle
@LeonieFierz, with two of the reviewers accepting your submission, let's move on.

There are a few tiny issues in the paper that need fixes:

- [ ]  Is there a DOI for the first reference, Allan et al 2021, that you can add?
- [ ] The reference to MURRAY_ET_AL_2020 is broken.
- [ ] You may want to add a full stop to the third item in the list in the "Statement of need" section.

Once this is done could you please:
