+++ paper/paper.bib
@@ -67,6 +67,7 @@
   title = {Linked Data Ontology for Urban and National Scale Building Energy Modelling},
   booktitle = {eSim conference proceedings},
   year = {2021}
+  url = {https://www.dora.lib4ri.ch/empa/islandora/object/empa%3A28038}
 }
 
 @article{JOHARI_ET_AL_2020,
