+++ paper/paper.bib
@@ -67,7 +67,7 @@
   title = {Linked Data Ontology for Urban and National Scale Building Energy Modelling},
   booktitle = {eSim conference proceedings},
   year = {2021}
-  url = {https://www.dora.lib4ri.ch/empa/islandora/object/empa%3A28038}
+  url = {http://www.ibpsa.org/proceedings/eSimPapers/2021/Contribution_1148_final_a.pdf}
 }
 
 @article{JOHARI_ET_AL_2020,
