+++ paper/paper.bib
@@ -66,7 +66,7 @@
   author = {James Allan and Leonie Fierz and Andrew Bollinger and Kristina Orehounig},
   title = {Linked Data Ontology for Urban and National Scale Building Energy Modelling},
   booktitle = {eSim conference proceedings},
-  year = {2021}
+  year = {2021},
   url = {http://www.ibpsa.org/proceedings/eSimPapers/2021/Contribution_1148_final_a.pdf}
 }
 
