### 2022-10-04 3c078d3
Try to fix paper.bib
For MURRAY_ET_AL_2020
