+++ paper/paper.bib
@@ -71,28 +71,26 @@
 }
 
 @article{JOHARI_ET_AL_2020,
-title = {Urban building energy modeling: State of the art and future prospects},
-journal = {Renewable and Sustainable Energy Reviews},
-volume = {128},
-pages = {109902},
-year = {2020},
-issn = {1364-0321},
-doi = {https://doi.org/10.1016/j.rser.2020.109902},
-url = {https://www.sciencedirect.com/science/article/pii/S1364032120301933},
-author = {F. Johari and G. Peronato and P. Sadeghian and X. Zhao and J. Widén},
-keywords = {Urban building energy modeling, Urban energy planning, Bottom-up energy modeling, Building archetype, Energy simulation, Thermal zoning},
-abstract = {During recent years, urban building energy modeling has become known as a novel approach for identification, support and improvement of sustainable urban development initiatives and energy efficiency measures in cities. Urban building energy models draw the required information from the energy analysis of buildings in the urban context and suggest options for effective implementation of interventions. The growing interest in urban building energy models among researchers, urban designers and authorities has led to the development of a diversity of models and tools, evolving from physical to more advanced hybrid models. By critically analyzing the published research, this paper incorporates an updated overview of the field of urban building energy modeling and investigates possibilities, challenges and shortcomings, as well as an outlook for future improvements. The survey of previous studies identifies technical bottlenecks and legal barriers in access to data, systematic and inherent uncertainties as well as insufficient resources as the main obstacles. Furthermore, this study suggests that the main route to further improvements in urban building energy modeling is its integration with other urban models, such as climate and outdoor comfort models, energy system models and, in particular, mobility models.}
+  title = {Urban building energy modeling: State of the art and future prospects},
+  journal = {Renewable and Sustainable Energy Reviews},
+  volume = {128},
+  pages = {109902},
+  year = {2020},
+  issn = {1364-0321},
+  doi = {10.1016/j.rser.2020.109902},
+  url = {https://www.sciencedirect.com/science/article/pii/S1364032120301933},
+  author = {F. Johari and G. Peronato and P. Sadeghian and X. Zhao and J. Widen},
+  keywords = {Urban building energy modeling, Urban energy planning, Bottom-up energy modeling, Building archetype, Energy simulation, Thermal zoning},
 }
 
 @article{MURRAY_ET_AL_2020,
-title = {Optimal transformation strategies for buildings, neighbourhoods and districts to reach CO2 emission reduction targets},
-journal = {Energy and Buildings},
-volume = {207},
-pages = {109569},
-year = {2020},
-issn = {0378-7788},
-doi = {https://doi.org/10.1016/j.enbuild.2019.109569},
-url = {https://www.sciencedirect.com/science/article/pii/S0378778819319486},
-author = {Portia Murray and Julien Marquant and Mathias Niffeler and Georgios Mavromatidis and Kristina Orehounig},
-keywords = {}
+  title = {Optimal transformation strategies for buildings, neighbourhoods and districts to reach CO2 emission reduction targets},
+  journal = {Energy and Buildings},
+  volume = {207},
+  pages = {109569},
+  year = {2020},
+  issn = {0378-7788},
+  doi = {10.1016/j.enbuild.2019.109569},
+  url = {https://www.sciencedirect.com/science/article/pii/S0378778819319486},
+  author = {Portia Murray and Julien Marquant and Mathias Niffeler and Georgios Mavromatidis and Kristina Orehounig}
 }
