### 2022-10-05 #4261 @LeonieFierz
> @LeonieFierz, with two of the reviewers accepting your submission, let's move on.
>
> There are a few tiny issues in the paper that need fixes:
>
>     * [x]   Is there a DOI for the first reference, Allan et al 2021, that you can add?
>
>     * [ ]  The reference to MURRAY_ET_AL_2020 is broken. => this was just on the joss_paper branch
>
>     * [x]  You may want to add a full stop to the third item in the list in the "Statement of need" section.
>
>
> Once this is done could you please:
>
>     * [x]  Make a tagged release of your software, and list the version tag of the archived version here.
>
>     * [x]  Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
>
>     * [x]  Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
>
>     * [x]  Please list the DOI of the archived version here.
>
>
> I can then move forward with accepting the submission.

@olejandro we tackled the points above

Tag for version 2.3.1: https://github.com/hues-platform/cesar-p-core/releases/tag/CESAR-P-V2.3.1
Zenoodo Archive DOI: 10.5281/zenodo.7150137

For your information, as there are no code changes compared to the version 2.3.0 we did not deploy an update to PyPi.

Is the joss-paper branch still needed or can we delete it as everything is now merged on master?
