### 2022-10-07 #4261 @timtroendle
@LeonieFierz Thanks! As you can see I do not manage to generate the paper anymore on master branch, therefore I suggest you to wait until this is resolved. There is only one minor issue with the archive: It lists an unknown author `boaa-empa`. Please make sure the list of authors matches the one on the paper.
