### 2022-10-07 #4261 @danielskatz
The error (in the error link above) is
> 'failure': Downloading of the repository failed. Please make sure the URL is correct. (Theoj::Error)

which makes me think there's something wrong with the repo name or branch name or permissions
