### 2022-10-07 #4261 @timtroendle
Good catch. I can confirm that cloning the repo doesn't work. Thank you @xuanxu and @danielskatz .

@LeonieFierz could you have a look at that so we can generate the pdf again?
