### 2022-10-11 #4261 @LeonieFierz
> @LeonieFierz Thanks! As you can see I do not manage to generate the paper anymore on master branch, therefore I suggest you to wait until this is resolved. There is only one minor issue with the archive: It lists an unknown author `boaa-empa`. Please make sure the list of authors matches the one on the paper.

@jasondegraw
I moved Aaron Bojarski (boaa-empa) to a Project Member, as he started working on the CESAR-P after we decided to go for the JOSS-Publication.
We added a data pack, so GitLFS has enought capacity and I just see that generating the paper works again.
