### 2022-10-11 #4261 @timtroendle
Thanks for the clarification, @LeonieFierz. Still, please make sure the author list on the paper and the repository are identical. If Aaron Bojarksi contributed to version v2.3.1 of CESAR-P, they should be listed on the paper as well.
