### 2022-10-12 #4261 @kthyng
Thank you for confirmation @olejandro! Want to make sure everything is clear before wrapping up.

@kristina-o A few comments on your paper:
* last sentence of first section: there is a colon after Zenodo. Seems like you either want no colon and keep the parenthetical reference or keep the colon and make the reference inline.
* Go through your references in detail to preserve capitalization and fix other small errors. You can preserve capitalization from your .bib file by placing {} around characters and words.
