+++ paper/paper.md
@@ -37,13 +37,13 @@ reducing energy consumption and integrating renewable energy sources in building
 of energy consumption patterns is needed at different temporal and spatial scales. CESAR-P (Combined Energy Simulation
 And Retrofitting - Python) is a bottom-up building stock modelling software to evaluate energy consumption and
 retrofitting strategies for individual buildings and neighbourhoods. CESAR-P is used to parameterize models for the
-dynamic building energy simulation tool EnergyPlus (@CRAWLEY_ET_AL_2001). It computes hourly energy demand profiles and
+dynamic building energy simulation tool EnergyPlus [@CRAWLEY_ET_AL_2001]. It computes hourly energy demand profiles and
 indoor temperatures at the building level and takes into account shading and reflections of surrounding buildings. After
 the current energy demand is computed, retrofitting measures for individual buildings can be evaluated in terms of
 energy savings, embodied emissions and costs. The default library of CESAR-P is based on Swiss archetypal buildings
 created from standards and statistics about the Swiss building stock; however, users are able to define their own
 constructions and internal conditions for their buildings. CESAR-P is written in Python and is a further development of
-the tool CESAR (@WANG_ET_AL_2018). The building model generator of CESAR-P creates EnergyPlus input data files (.IDF)
+the tool CESAR [@WANG_ET_AL_2018]. The building model generator of CESAR-P creates EnergyPlus input data files (.IDF)
 for each building within a specified extent, and can process geo-referenced input data as shapefiles (ESRI .shp format)
 for geometry. Building usage type and construction information is used to populate the building model. The EnergyPlus
 files are executed with a weather file (EnergyPlus epw weather data format) of the geographic location which provides
@@ -81,7 +81,7 @@ stored in archetypes. The target audience of the software are energy planners, e
   shapefiles), simplifying the generation of required input files.
 - The archetypes used to parameterize the building models are stored as RDF triples (in TTL files). The triples contain
   ABox statements related to the construction properties defined by each archetype. The statements conform to classes
-  and properties stored in the urban energy simulation ontology (@ALLAN_ET_AL_2021). By default, a construction
+  and properties stored in the urban energy simulation ontology [@ALLAN_ET_AL_2021]. By default, a construction
   archetype will be assigned to a building instance using the building age. The construction archetype contains all of
   the material layers in the floor, roof and external walls of the building. CESAR-P loads the archetype data directly
   from the provided TTL files, but it can be configured to access a graph database instance to query archetype data. The
@@ -91,12 +91,12 @@ stored in archetypes. The target audience of the software are energy planners, e
   enable the linking of concepts and data models used in other studies using TBox inferences.
 - Evaluation of retrofitting options and strategies.
 - Simulation and comparison of energy demand for scenarios with different building parametrization, for example to
-  evaluate passive cooling potential (@SILVA_ET_AL_2022).
+  evaluate passive cooling potential [@SILVA_ET_AL_2022].
 - Since CESAR-P uses EnergyPlus, which is a continually developed energy simulation tool within the research field, the
   capabilities of CESAR-P can be extended to accommodate future developments and improve modelling accuracy.
 - The tool can be applied at various scales and can be used for predicting the energy consumption of a single building,
-  of whole neighbourhoods with thousands of buildings (@WANG_ET_AL_2018) or, in combination with clustering approaches,
-  for calculating national building energy demands (@MURRAY_ET_AL_2020, @EGGIMANN_ET_AL_2022).
+  of whole neighbourhoods with thousands of buildings ([@WANG_ET_AL_2018] or, in combination with clustering approaches,
+  for calculating national building energy demands [@MURRAY_ET_AL_2020, @EGGIMANN_ET_AL_2022].
 
 # Example
 
