+++ README.rst
@@ -109,9 +109,9 @@ The base methodology of CESAR-P regarding building simulation and retrofit is se
 
 * To cite the CESAR-P OpenSource version:
 
-  Leonie Fierz, Urban Energy Systems Lab, Empa. (2021, April 13). hues-platform/cesar-p-core: 1.3.0 (Version 1.3.0).  Zenodo. http://doi.org/10.5281/zenodo.4682881
-  
-  Leonie Fierz, Urban Energy Systems Lab, Empa. (2021, July 30). hues-platform/cesar-p-core: CESAR-P-V2.0.1 (CESAR-P-V2.0.1). Zenodo. https://doi.org/10.5281/zenodo.5148531
+
+  Léonie Fierz, Aaron Bojarski, Urban Energy Systems Lab, Empa. (2022)
+  hues-platform/cesar-p-core. Zenodo. https://doi.org/10.5281/zenodo.4682880
 
 
 Installation & Usage
