### 2022-10-13 0cf61bb
Update contributors and reference to zenoodo in paper
