+++ paper/paper.bib
@@ -53,7 +53,7 @@
 }
 
 @software{cesar-p-zenodo,
-  author = {Léonie Fierz and UrbanEnergySystemsLab},
+  author = {Léonie Fierz and Aaron Bojarski and UrbanEnergySystemsLab},
   title = {hues-platform/cesar-p-core},
   month = {apr},
   year = {2021},
