### 2022-10-13 a38a73b
some grammer improvements for the paper text
