+++ paper/paper.md
@@ -49,29 +49,14 @@ for geometry. Building usage type and construction information is used to popula
 files are executed with a weather file (EnergyPlus epw weather data format) of the geographic location which provides
 the necessary climatic context for the building. The results consist of a configurable set of yearly demand values and
 hourly resolved times series for each building from EnergyPlus and additional operational costs and emissions. The
-retrofit module can be used to update the building models according to a specified retrofit strategy. Retrofit
-strategies specify the frequency of retrofits to the main building elements such as walls, windows and roofs across the
-building stock. The output of the retrofit module include, in addition to the above-mentioned operational indicators,
-costs and embodied emissions of retrofitting interventions. All simulation steps can also be run in parallel on multiple
-cores. The code is platform-independent and is tested to run on Linux and Windows. The source code for CESAR-P has been
-archived to Zenodo [@cesar-p-zenodo] 
+retrofit module can be used to modify the parameters of individual building models according to a specified retrofit strategy. Retrofit strategies specify the frequency of retrofits to the main building elements such as walls, windows and roofs across the building stock. The output of the retrofit module include, in addition to the above-mentioned operational indicators, costs and embodied emissions of retrofitting interventions. All simulation steps can also be run in parallel on multiple cores. The code is platform-independent and is tested to run on Linux and Windows. The source code for CESAR-P has been archived to Zenodo [@cesar-p-zenodo] 
 
 # Statement of need and Key features 
 
 Building energy simulation is used to generate the expected energy demand profiles of buildings. Building energy demand
 profiles are useful for planners to evaluate current performance and investigate how buildings will perform under
-different scenarios (e.g. building retrofits, changing climate). Measured demands are often unavailable which makes
-building energy simulation a key tool for decision-making. Urban building energy modelling (UBEM) has been established
-in recent years to support the implementation of energy conservation measures and the integration of renewable energy
-technologies at city and district scale. UBEM tools can be classified as taking a top-down or a bottom-up modelling
-approach. Top-down approaches rely on aggregated data and different drivers (e.g. socio-economic indicators or climate
-data), whereas bottom-up modelling approaches are based on data at an individual building scale and include either
-statistical or engineering based modelling techniques to compute energy demand data of a set of individual buildings
-[@JOHARI_ET_AL_2020]. Various review papers summarize available tools, including their key-features and limitations
-[@JOHARI_ET_AL_2020]. The here presented tool CESAR-P falls within the category of bottom-up modelling approaches. CESAR-P
-uses EnergyPlus as the simulation engine, an open-source building energy simulation software that is widely-used in both
-academia and industry. CESAR-P simplifies the creation of EnergyPlus input files using a set of key parameters which are
-stored in archetypes. The target audience of the software are energy planners, energy utilities, cities and researchers. The time-resolved energy consumption of buildings at large scale that CESAR-P simulates are crucial for developing renewable energy integration and retrofit strategies at different scales, from the building to the neighborhood and city scale. Key features of CESAR-P include: 
[…]
