### 2022-10-13 af19fc1
fix capitalization in paper.bib
