+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{WANG_ET_AL_2018,
-  title = {CESAR: A bottom-up building stock modelling tool for Swit-zerland to address sustainable energy transformation strategies},
+  title = {CESAR: A bottom-up building stock modelling tool for {Switzerland} to address sustainable energy transformation strategies},
   journal = {Energy and Buildings},
   volume = {169},
   pages = {9-26},
@@ -40,7 +40,7 @@
 }
 
 @article{SILVA_ET_AL_2022,
-  title = {Opportunities for passive cooling to mitigate the impact of climate change in Switzerland},
+  title = {Opportunities for passive cooling to mitigate the impact of climate change in {Switzerland}},
   journal = {Buildings and environment},
   volume = {},
   pages = {108574},
