### 2022-10-14 007ac58
Update paper.md
Additional corrections to the text
