+++ paper/paper.md
@@ -41,7 +41,7 @@ dynamic building energy simulation tool EnergyPlus [@CRAWLEY_ET_AL_2001]. It com
 indoor temperatures at the building level and takes into account shading and reflections of surrounding buildings. After
 the current energy demand is computed, retrofitting measures for individual buildings can be evaluated in terms of
 energy savings, embodied emissions and costs. The default library of CESAR-P is based on Swiss archetypal buildings
-created from standards and statistics about the Swiss building stock; however, users are able to define their own
+created from standards and statistics about the Swiss building stock; however, users can define their custom
 constructions and internal conditions for their buildings. CESAR-P is written in Python and is a further development of
 the tool CESAR [@WANG_ET_AL_2018]. The building model generator of CESAR-P creates EnergyPlus input data files (.IDF)
 for each building within a specified extent, and can process geo-referenced input data as shapefiles (ESRI .shp format)
@@ -49,20 +49,20 @@ for geometry. Building usage type and construction information is used to popula
 files are executed with a weather file (EnergyPlus epw weather data format) of the geographic location which provides
 the necessary climatic context for the building. The results consist of a configurable set of yearly demand values and
 hourly resolved times series for each building from EnergyPlus and additional operational costs and emissions. The
-retrofit module can be used to modify the parameters of individual building models according to a specified retrofit strategy. Retrofit strategies specify the frequency of retrofits to the main building elements such as walls, windows and roofs across the building stock. The output of the retrofit module include, in addition to the above-mentioned operational indicators, costs and embodied emissions of retrofitting interventions. All simulation steps can also be run in parallel on multiple cores. The code is platform-independent and is tested to run on Linux and Windows. The source code for CESAR-P has been archived to Zenodo [@cesar-p-zenodo] 
+retrofit module can be used to modify the parameters of individual building models according to a specified retrofit strategy. Retrofit strategies specify the frequency of retrofits to the main building elements such as walls, windows and roofs across the building stock. The output of the retrofit module includes, in addition to the above-mentioned operational indicators, costs and embodied emissions of retrofitting interventions. All simulation steps can also be run in parallel on multiple cores. The code is platform-independent and is tested to run on Linux and Windows. The source code for CESAR-P has been archived to Zenodo [@cesar-p-zenodo] 
 
 # Statement of need and Key features 
 
 Building energy simulation is used to generate the expected energy demand profiles of buildings. Building energy demand
 profiles are useful for planners to evaluate current performance and investigate how buildings will perform under
[…]
